+++ server.js
@@ -130,6 +130,7 @@ const MIME_TYPES = {
   ".txt": "text/plain; charset=utf-8",
   ".woff": "font/woff",
   ".woff2": "font/woff2",
+  ".webmanifest": "application/manifest+json; charset=utf-8",
 };
 
 const server = http.createServer((req, res) => {